requirements.txt
fastapi
uvicorn[standard]
elasticsearch
python-dotenv
pydantic
requests
sendgrid
schedule